Test: Phase A3 – Split UniversalCreateService and UniversalSearchService\n\n- Split UniversalCreateService into core-resources / objects-deals-tasks / validation-errors\n- Split UniversalSearchService into companies-people-lists / records-tasks\n- Update test-impact-mapping.json\n\nRefs #534 (Phase A3)
